admin_events_attendees/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic[email]==1.10.13
python-jose[cryptography]==3.5.0
bcrypt==4.0.1
python-multipart==0.0.6
python-decouple==3.8
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pyotp==2.9.0
qrcode[pil]==7.4.2
pillow==10.1.0
